Further generalized the concept of entities and refactored entity image context
