+++ pupsquad/__init__.py
@@ -2,9 +2,9 @@
 import numpy as np
 import pygame
 
-from pupsquad import decor
-from pupsquad import entity
-from pupsquad import scene
+import pupsquad.decor
+import pupsquad.scene
+import pupsquad.player
 from pupsquad.constant import METERS
 from pupsquad.constant import SCREEN_WIDTH
 from pupsquad.constant import SCREEN_HEIGHT
@@ -12,14 +12,15 @@ from pupsquad.constant import SCREEN_HEIGHT
 
 def main():
     """Main execution function for running the application."""
+    # Initialize some PyGame stuff.
     pygame.init()
     screen = pygame.display.set_mode((SCREEN_WIDTH, SCREEN_HEIGHT))
-    # Settings for the first level.
+    # Compose the first level.
     # In time this should all be replaced by a file that is read and parsed
     # entirely by the 'Level' class.
     ## The player.
     player_position = np.array([2.*METERS, 5.8*METERS])
-    player = entity.Player(player_position)
+    player_ = pupsquad.player.Player(player_position)
     ## The decor.
     grid = np.array([
             [1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1],
@@ -53,9 +54,10 @@ def main():
             [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
             [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
         ])
-    level_decor = decor.Decor(grid)
-    level = scene.Level(player, level_decor)
-    game = scene.SceneContext(level)
+    level_decor = pupsquad.decor.Decor(grid)
+    level = pupsquad.scene.Level(player_, level_decor)
+    # Initialize and run the game.
+    game = pupsquad.scene.SceneContext(level)
     game.run(screen)
